Transfer sine generator from tutorial 1 into agents' module
